2 more apps: AEBStream -> Multiple Streams : burst-> wait -> cont rate -> burst ->... AEData -> Multiple Streams : Gen data -> send cont rate->wait -> gen data -> ...
Each stream config via xml (bitrate, waiting time, burst prob, etc.)

+++ src/DAF/AE/AEVoice/AEVoice.cc
@@ -69,7 +69,7 @@ int Stream::getid(){
 }
 
 void Stream::compute(){
-    packetSize = ceil(brt*wt);
+    packetSize = (int) ceil(brt*wt);
     minSTime = avgSTime-varSTime;
     maxSTime = avgSTime+varSTime;
     minWTime = avgWTime-varSTime;
